polish

+++ core/libs/threadimageio/fileio/loadingcache.cpp
@@ -180,8 +180,11 @@ LoadingCache::LoadingCache()
     : d(new Private(this))
 {
     DMemoryInfo memory;
-    setCacheSize(qBound(quint64(100), quint64(memory.totalPhysical() / 1024 / 1024 * 0.06), quint64(1024)));
-    setThumbnailCacheSize(10, 200); // the pixmap number should not be based on system memory, it's graphics memory
+    setCacheSize(qBound(100, int(memory.totalPhysical() / 1024 / 1024 * 0.06), 1024));
+
+    // the pixmap number should not be based on system memory, it's graphics memory
+
+    setThumbnailCacheSize(10, 200);
 
     // good place to call it here as LoadingCache is a singleton
